refactor: move index template into svgr config
