--- svgr.index-template.js
@@ -1,12 +1,0 @@
-const path = require('path');
-
-function defaultIndexTemplate(filePaths) {
-  const exportEntries = filePaths.map((filePath) => {
-    const basename = path.basename(filePath, path.extname(filePath));
-    const exportName = `${basename}SvgIcon`;
-    return `export { default as ${exportName} } from './${basename}'`;
-  });
-  return exportEntries.join('\n');
-}
-
-module.exports = defaultIndexTemplate;
